database: respect "workflow.public" column
It's possible for a workflow to be public _and_ have a secret_id. In
that case, public=true means the workflow is allowed.

We haven't defined what should happen when public=true but the user
enters a wrong token. For now, the request will succeed. In the future,
we may change this.